update: remove some redundant chart state & pumb version oraidex-common-ui
