+++ src/pages/UniversalSwap/index.module.scss
@@ -8,7 +8,7 @@
 
   .tv-chart {
     position: relative;
-    height: 420px;
+    height: 500px;
     padding-top: 10px;
     transition: all 1s ease;
